refactor(expose): move the currency map out of the registry client
The registry client had two jobs: talk to the node, and know which hex
unit string names USDM on each network. The second one is self contained
and was in the way of the file size ceiling, so it moves to its own
module. registry re-exports the names so callers do not change.
